% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap-latent-space.R
\name{ASE}
\alias{ASE}
\title{Adjacency Spectral Embedding}
\usage{
ASE(network, d)
}
\arguments{
\item{network}{An igraph object with \eqn{n} nodes.}

\item{d}{number of latent dimensions to embed the node positions in.}
}
\value{
An \eqn{n}-by-\eqn{d} matrix where the \eqn{i^{th}} row of the matrix corresponds to node \eqn{i}'s position in the \eqn{d}-dimensional latent space.
Bootstrap samples will retain vertex names of the input \code{network} but will loose all other vertex and edge attributes.
}
\description{
Calculate the adjacency spectral embedding of a network.
}
\details{
This calculates the latent positions of the nodes of a network in \eqn{\mathbb{R}^d}.

ADD ALGORITHM
}
\examples{
#ADD EXAMPLE ONCE DATA IS ADDED
}
\references{
\insertRef{sussman-et-al-2012}{JaB}
}
