% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-local.R
\name{bootstrap_local}
\alias{bootstrap_local}
\title{Local Bootstrap}
\usage{
bootstrap_local(
  network,
  B = 3000,
  quantile_n = NULL,
  dist.func = dist_ASE,
  weighted = FALSE,
  fast = NULL,
  node.names = NULL
)
}
\arguments{
\item{network}{An \code{igraph} object with \eqn{n} nodes.}

\item{B}{number of bootstrap samples to calculate}

\item{quantile_n}{Parameter for \link[localboot:localboot]{localboot} function. The quantile used for neighborhood selection in the \link[localboot:localboot]{localboot} function. Defaults to \code{NULL} which uses \eqn{\sqrt{log(n)/n}}.}

\item{dist.func}{Parameter for \link[localboot:localboot]{localboot} function. The distance function used to calculate distances between nodes in \code{network}. Defaults to \code{get_dist_default_eigen}, an internal function in the \pkg{localboot} package.}

\item{weighted}{Parameter for \link[localboot:localboot]{localboot} function. A logical indicating if the network is weighted. Defaults to \code{FALSE}.}

\item{fast}{Parameter for \link[localboot:localboot]{localboot} function. A logical indicating if a faster, approximate method should be used. Automatically set based on network size if NULL.}
}
\value{
A list of length \code{B} where each element is a bootstrapped network as an igraph object.
Each bootstrapped network, \eqn{G^{(b)}} has the following vertex attributes:
\itemize{
\item \code{boot.index} is the index of the node in the bootstrap sample as \eqn{vb1, vb2, ..., vbn}.
\item \code{original.index.cpp} this is the index of the corresponding node in
the original network as \eqn{v0, v1, ..., v(n-1)}.  The \link[localboot:localboot]{localboot}
function uses C++ which is the reason for this index.
For example if \eqn{v_1^{b)}} is the \eqn{4^{th}} node in the original \code{network}, the \code{original.index.cpp[1]} is \eqn{v3}.
\item \code{original.index.r}this is the index of the corresponding node in
the original network as \eqn{v1, v2, ..., vn}.  The \link[localboot:localboot]{localboot}
function uses C++, but we want to use R's indexing from 1 to n.
For example if \eqn{v_1^{b)}} is the \eqn{4^{th}} node in the original \code{network}, then \code{original.index.cpp[1]} is \eqn{v4}.
\item \code{name} is the name of the corresponding node in the original network.
For example if \eqn{v_1^{b)}} is the \eqn{4^{th}} node in the original \code{network},
then \code{name[1]} is \code{V(network)$name[4]}.
}
}
\description{
Generate bootstrap samples of a network with the local bootstrap method described in \insertCite{zu-quin-2025;textual}{JaB}
from the \pkg{localboot} package.
}
\details{
This is a wrapper function to generate bootstrap samples of a network with
the local bootstrap method described in \insertCite{zu-quin-2025;textual}{JaB}.
This function uses the \link[localboot:localboot]{localboot} function to
generate bootstrap samples, then formats them to fit in the JaB package syntax.

Let \eqn{\boldsymbol{A}} be the adjacency matrix of \code{network} \eqn{G } with \eqn{n} nodes
in node set \eqn{V}.

Let \eqn{\boldsymbol{D}} be the distance matrix for all pairs of nodes where \eqn{d_{i,j}}
is the distance between node \eqn{i} and node \eqn{i} according to the distance function \code{dist.func}.

Construct a neighborhood for each node as \eqn{\mathcal{N}(v_i) = \{v_j \in V\ : d_{i,j} < d_i^{(s)} \}}
where \eqn{d_i^{(s)} } is the \eqn{s^{th}} smallest distance of \eqn{d_{i,1}, ..., d_{i, n}}.
This construction ensures each neighborhood always has \eqn{s = |\mathcal(N)(v_i)|} neighbors.

To generate a bootstrap adjacency matrix \eqn{\boldsymbol{A}^{(b)}}, for
\eqn{1 \leq i < j \leq n} independently draw
\deqn{
\begin{align}
A_{ij}^{(b)} &\sim \text{Uniform} \Big(\{A_{k\ell} : k \in \mathcal{N}(v_{i}^{(b)}), \ell \in \mathcal{N}(v_{j}^{(b)})\} \Big) \\
v_i^{(b)} &\sim \text{Uniform}(V)
\end{align}
}

There may be cases in which \eqn{\mathcal{N}(v_{i}^{(b)})} and \eqn{\mathcal{N}(v_{j}^{(b)})}
are overlapping sets. In these cases, set  \eqn{\mathcal{N}(v_{i}^{(b)}) \cap \mathcal{N}(v_{j}^{(b)}) = \emptyset}.
}
\examples{
library(JaB)
library(igraph)
library(localboot)
data("paul.revere")

# generate 10 bootstrap samples
bootstrap_local(paul.revere, B = 10)

# Use a new distance function
}
\references{
\insertAllCited{}
}
