% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{detect_type}
\alias{detect_type}
\title{Detect the Type of Objects in a List}
\usage{
detect_type(obj)
}
\arguments{
\item{obj}{A list of objects to be analyzed. All objects in the list must belong to the same class.}
}
\value{
A character string indicating the type of objects in the list. Possible return values are:
\itemize{
\item \code{"igraph"}: If the objects are igraph objects.
\item \code{"dgCMatrix"}: If the objects are sparse matrices of class \code{dgCMatrix}.
\item \code{"edgelist"}: If the objects are matrices with two columns, representing an edge list.
\item \code{"matrix"}: If the objects are square matrices, typically representing adjacency matrices.
}
}
\description{
Attempts to detect the type of objects in a list based on their class and dimensions, as used in the JaB package. This function checks if all elements in the list are of the same class and determines their type, such as \code{"igraph"}, \code{"dgCMatrix"}, \code{"edgelist"}, or \code{"matrix"}.
}
\details{
The function first verifies that all objects in the list are of the same class. It then checks if the class matches known types (\code{"igraph"} or \code{"dgCMatrix"}). If the class is not one of these, it further examines the dimensions of each object to determine if they represent an edge list (two-column matrix) or an adjacency matrix (square matrix).
}
