% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-local-boot.R
\name{dist_ASE}
\alias{dist_ASE}
\title{Euclidean distance of adjacency spectral embeddings}
\usage{
dist_ASE(A, d = 2)
}
\arguments{
\item{A}{An \eqn{n \times n} adjacency matrix}

\item{d}{dimension of the latent space embed in}
}
\value{
An \eqn{n \times n} matrix \eqn{D}, where \eqn{D_{i,j}} represents the distance
between nodes \eqn{v_i} and \eqn{v_j}.
}
\description{
Compute the Euclidean distance of adjacency spectral embeddings
}
\details{
For adjacency matrix \eqn{\boldsymbol{A}} with \eqn{n} nodes,
let \eqn{\hat{\boldsymbol{\Lambda}} \in \mathbb{R}^{d \times d}}
be the diagonal matrix formed by the top \code{d} largest-magnitude eigenvalues
of the adjacency matrix and \eqn{\hat{\boldsymbol{U}} \in \mathbb{R}^{n \times d}}
be the matrix with the corresponding eigenvectors as its columns.
The adjacency spectral embedding of \eqn{\boldsymbol{A}} is
\eqn{\hat{\boldsymbol{X}} = \hat{\boldsymbol{U}}\hat{\boldsymbol{\Lambda}}^{1/2} \in \mathbb{R}^{n \times d}}.

Then, \eqn{D_{i,j}} is the Euclidean norm of \eqn{\hat{\boldsymbol{X}}_{i, \cdot}} and \eqn{\hat{\boldsymbol{X}}_{j, \cdot}}.
}
\examples{
library(JaB)
library(igraph)
data("paul.revere")

A <- as_adjacency_matrix(paul.revere, sparse = FALSE)
D <- dist_ASE(A)

}
