% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-local-boot.R
\name{dist_max}
\alias{dist_max}
\title{Upper Bound distance}
\usage{
dist_max(A)
}
\arguments{
\item{A}{An \eqn{n \times n} adjacency matrix}
}
\value{
An \eqn{n \times n} matrix \eqn{D}, where \eqn{D_{i,j}} represents the computed distance
between nodes \eqn{v_i} and \eqn{v_j}.
}
\description{
Compute the upper bound distance defined in \insertCite{zhang-levina-zhu-2017;textual}{JaB}.
}
\details{
Computes distance matrix \eqn{D} where each
\eqn{d_{i,j} = \max_{k\neq i,j}| \langle A_{i, \cdot} - A_{j, \cdot} , A_{k, \cdot} \rangle | /n}

\eqn{d_{i,j}} is computed by \code{dist_max_ij} internal function.
}
\examples{
library(JaB)
library(igraph)
data("paul.revere")

A <- as_adjacency_matrix(paul.revere, sparse = FALSE)
D <- dist_max(A)

}
\references{
\insertAllCited{}
}
