% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{ess}
\alias{ess}
\title{Calculate effective sample size for JAB}
\usage{
ess(n1, n2 = NULL)
}
\arguments{
\item{n1}{Numeric. One or more sample sizes of the first sample.}

\item{n2}{Numeric. One or more sample sizes of the second sample. If
\code{NULL}, a one-sample t-test is assumed and the function returns \code{n1}.}
}
\value{
One or more effective sample sizes to be used in the calculation
of JAB.
}
\description{
For one- and two-sample t-tests this function determines the effective
N to be used in the JAB calculation.
}
\details{
For the two-sample t-test, the effective sample size $N$ is
derived from the formula of the standard error and thus is
$(n_1 \cdot n_2) / (n_1 + n_2)$.
}
\examples{
sleep_ttest <- t.test(extra ~ group, data = sleep, var.equal = TRUE)
n <- table(sleep$group)
jab_p(sleep_ttest$p.value, n = ess(n1 = n[1], n[2]))
}
