% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{ess_3pn}
\alias{ess_3pn}
\title{Calculate required effective sample size}
\usage{
ess_3pn(x, p)
}
\arguments{
\item{x}{Numeric. Bayes factor in favor of the null hypothesis.}

\item{p}{Numeric. Probability of obtaining a Bayes factor of \code{x} or larger.}
}
\value{
The effective sample size for the given \code{x} and \code{p}.
}
\description{
This function calculates the effective sample size required to achieve a
Bayes factor with a defined probability according to the $3p\sqrt{n}-rule
\emph{assuming the null hypothesis is true}.
}
\details{
The function uses different formulas based on the value of \code{x} to calculate the effective sample size.
}
\examples{
ess_3pn(0.05, 0.5)
ess_3pn(0.3, 0.7)
}
