% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-centrality.R
\name{get_centrality}
\alias{get_centrality}
\title{Calculate give centrality statistic on a network}
\usage{
get_centrality(network, func.name, package.name = NULL, func.args = NULL)
}
\arguments{
\item{network}{The network object. This should be the appropriate object type needed for the `func.name` specified.}

\item{func.name}{A character string of the function name used to calculate the desired centrality statistic.}

\item{package.name}{A character string of the name of the package the `func.name` function is in. If left as `NULL` if the function will be called as loaded in the users environment.}

\item{func.args}{A list of additional arguments the `func.name` function may need beyond the `network` object.}
}
\value{
A numeric vector of the centrality statistic as calculated by `func.name`.
}
\description{
Calculate give centrality statistic on a network
}
\examples{
#COME BACK TO THESE EXAMPLES ONCE DATA IS WRITTEN
#get_centrality("closeness", paul.igraph, package.name = "igraph, func.args = list(normalized = T))
#get_centrality("degree", paul.igraph, package.name = "igraph, func.args = list(normalized = T))
}
