% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-largest-component.R
\name{get_largest_component}
\alias{get_largest_component}
\title{Find Largest Component of a Network}
\usage{
get_largest_component(network)
}
\arguments{
\item{network}{An \code{igraph} object.}
}
\value{
An \code{igraph} object which is the largest connected component of the network.
If the network is already connected, the function returns the input network unchanged.
}
\description{
This function takes in an \code{igraph} network object and returns and \code{igraph} object
of the largest connected component.
}
\examples{
library(igraph)
#unconnected network
data("hamster")
is.connected(hamster)
get_largest_component(hamster)

#connected network
data("karate")
is.connected(karate)
get_largest_component(karate)
}
