% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{get_nodes}
\alias{get_nodes}
\title{Node Names from a Network Object}
\usage{
get_nodes(network)
}
\arguments{
\item{network}{A network object with \eqn{n}, which may be of types \code{matrix}, \code{dgCMatrix} \code{edgelist}, or \code{igraph}).}
}
\value{
A vector of node names. If node names are not assigned, it returns a sequence of integers from 1 to the number of nodes.
}
\description{
Extracts the names of nodes (vertices) from a network object.
This is for internal use for network types in the JaB package
}
\seealso{
\code{\link{detect_type}}, \code{\link{make_network_type}}
}
