% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamster}
\alias{hamster}
\title{Pester Hamster Network}
\format{
An undirected \code{igraph} graph object. Nodes are individual accounts from hamsterster.com
and an edge is present between two nodes if the respective accounts are friends.
}
\source{
\url{https://networkrepository.com/petster-hamster.php}
}
\usage{
hamster
}
\description{
Friendship network among users of hamsterster.com from \insertCite{rossi-ahmed-2015;textual}{JaB}
}
\examples{
data("hamster")
set.seed(600)
plot(hamster, vertex.label = NA, vertex.size = 3)

}
\references{
\insertAllCited{}
}
