% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jab.R
\name{jab}
\alias{jab}
\alias{jab.default}
\alias{jab.htest}
\alias{jab.lm}
\alias{jab.glht}
\alias{jab.cocor.dep.groups.overlap}
\alias{jab.merMod}
\alias{jab.glmerMod}
\alias{jab.emmGrid}
\title{Jeffrey's approximate Bayes factor}
\usage{
jab(x, ...)

\method{jab}{default}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{htest}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{lm}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{glht}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{cocor.dep.groups.overlap}(x, test, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{merMod}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{glmerMod}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))

\method{jab}{emmGrid}(x, prior, ..., curvature = NULL, ratio = getOption("jab.ratio"))
}
\arguments{
\item{x}{Frequentist analysis output}

\item{...}{Additional arguements to pass to \code{prior}.}

\item{prior}{Function. Function to compute prior density at MLE.}

\item{curvature}{Function. Function to compute curvature at MLE. Must accept
the same arguments as \code{prior} as the ellipsis is passed to both functions.}

\item{ratio}{Character. Either \code{"01"} or \code{"10"} to define the
direction of the evidence ratio.}

\item{test}{Character. Wald statistics to use for calculations.}
}
\value{
A numeric (vector of) Jeffrey's approximate Bayes factors.
}
\description{
Generic function to compute Jeffrey's approximate Bayes factors for
results from frequentist Wald tests.
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)

lm.D9 <- lm(weight ~ group)
jab(lm.D9, prior = dcauchy, location = 0, scale = sqrt(2)/4)
}
