% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jab.R
\name{jab_3pn}
\alias{jab_3pn}
\title{Jeffrey's approximate Bayes factor}
\usage{
jab_3pn(x, n, ratio = getOption("jab.ratio"))
}
\arguments{
\item{x}{Numeric. \emph{p}-value from frequentist Wald test.}

\item{n}{Numeric. Sample size.}

\item{ratio}{Character. Either \code{"01"} or \code{"10"} to define the
direction of the evidence ratio.}
}
\value{
A numeric (vector of) Jeffrey's approximate Bayes factors.
}
\description{
Piecewise approximation of Jeffrey's approximate Bayes factors for
results from frequentist Wald tests form p-values.
}
\details{
Bayes factors are calculated using the precise piecewise
approximation method suggested by
\insertCite{Wagenmakers2022;textual}{jab},

\deqn{
   \text{JAB}_{01} = \begin{cases}
   3p \sqrt{n} \text{} & \text{if }p \leq .10 \\
   \frac{4}{3}p^{2/3} \sqrt{n} \text{} & \text{if }0.1 < p \leq .50 \\
   3p \sqrt{n} \text{} & \text{if }p > .50
   \end{cases}
   }
}
\examples{
jab_3pn(x = 0.0495, n = 136)
}
\references{
\insertAllCited{}
}
