% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jab-lm.R
\name{jab_lm}
\alias{jab_lm}
\title{Jackknife-after-Bootstrap for Linear Regression}
\usage{
jab_lm(
  mod,
  stat = "rstudent",
  quant.lower = 0.05,
  quant.upper = 0.95,
  B = 3100,
  package.name = NULL,
  stat.args = NULL
)
}
\arguments{
\item{mod}{An \code{lm} object. Output from \link{lm}{stats}.}

\item{stat}{A character string of the function name used to calculate the desired centrality statistic.}

\item{quant.lower}{A numeric between 0 and 1 used as the lower cutoff in the JaB algorithm. Default is 0.05. Must be smaller than \code{quant.upper}.}

\item{quant.upper}{A numeric between 0 and 1 used as the upper cutoff in the JaB algorithm. Default is 0.95. Must be larger than \code{quant.lower}.}

\item{B}{Number of bootstrap samples. Default is 3100}

\item{package.name}{A character string of the name of the package that \code{func.name} function is in. If left as \code{NULL} if the function will be called as loaded in the users environment.}

\item{stat.args}{A named list of additional arguments the \code{func.name} function may need beyond the \code{mod} object.}
}
\value{
A data frame with 5 columns.
\itemize{
\item{"row.ID": }{Row Number of the observation in the data set used for \code{mod}.}
\item{"lower": }{Lower quantile cutoff as determined by \code{quant.lower}. }
\item{"upper": }{Upper quantile cutoff as determined by \code{quant.upper}. }
\item{"orig": }{The original influence/outlier statistic calculated by \code{stat}.}
\item{"influential": }{Logical flagging if the observation is influential or not. TRUE if \code{orig} < \code{lower} or \code{orig} > \code{upper}. FALSE otherwise.}
}
}
\description{
ADD DESCRIPTION - add something that the default is 0.05-0.95 i.e. center 90\%- why default for B =3100
ADD CITATIONS
}
\examples{
library(stats)
data("LifeCycleSavings")

mod <- lm(sr ~ ., data = LifeCycleSavings)

# JaB with DFFITS
result1 <- jab_lm(mod,
                  stat = "dffits",
                  quant.lower = 0.025,
                  quant.upper = 0.975,
                  B = 3100)
result1[result1$influential, ]


#define the likelihood distance as influence statistic
## NEED TO CITE
infl_like <- function(mod){

  n <- length(mod$fitted.values)
  p <- length(mod$coefficients)

  ti <- rstudent(mod)
  h <- hatvalues(mod)

  p1 <- log( (n/(n-1)) * ((n-p-1) / (ti^2 +n-p-1)) )
  p2 <- ti^2 * (n-1) / (1-h) / (n-p-1)

  return(n*p1 + p2 - 1)
}

## JaB with Likelihood Distance
# JaB with DFFITS
result2 <- jab_lm(mod,
                  stat = "infl_like",
                  quant.lower = 0.05,
                  quant.upper = 0.95,
                  B = 3100)
result2[result2$influential, ]
}
