% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bootstrap-centrality.R
\name{make_igraph}
\alias{make_igraph}
\title{Convert an Object to an igraph Graph}
\usage{
make_igraph(obj, obj.type, mode = "undirected", weighted = NULL)
}
\arguments{
\item{obj}{Object output from `bootstrap_` functions. Must be of type `igraph`, `matrix`, `sparsematrix`, or `edgelist`.}

\item{obj.type}{A character string specifying the type of `obj`. Options include `"igraph"`, `"matrix"`, `"sparsematrix"`, or `"edgelist"`. Same as `output.type` parameter in `bootstrap_` function.}

\item{mode}{A character string specifying whether the graph is `directed`, `undirected`, `max`, `min`, `upper`, `lower`, `plus`).
Defaults to `undirected`. See `igraph::graph_from_adjacency_matrix` for details.}

\item{weighted}{A logical indicating if the graph is weighted. If `NULL`, no weights are added. Defaults to `NULL`.}
}
\value{
An igraph object representing the graph.
}
\description{
Converts an object of specified type (e.g., adjacency matrix, edgelist) to an igraph graph object.
Supported input types include `igraph`, `matrix`, `sparsematrix`, and `edgelist`.
}
\examples{
# For an adjacency matrix
#adj_matrix <- matrix(c(0, 1, 1, 0), nrow = 2)
#make_igraph(adj_matrix, obj.type = "matrix")

# For an edge list
#edge_list <- matrix(c(1, 2, 2, 3), ncol = 2)
#make_igraph(edge_list, obj.type = "edgelist", mode = "directed")

}
