% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallows-distance-test.R
\name{mallow_equiv_test}
\alias{mallow_equiv_test}
\title{Equivalence Test of Two Vectors Using Mallows Distance}
\usage{
mallow_equiv_test(vec1, vec2, alpha, delta0)
}
\arguments{
\item{vec1}{A numeric vector.}

\item{vec2}{A numeric vector of the same length as \code{vec1}.}

\item{alpha}{Trimming parameter \eqn{\alpha \in(0, 0.5)} for trimmed Mallows' distance.}

\item{delta0}{Tolerance value for hypothesis test.}
}
\description{
Conduct the equivalence test using the \eqn{p^{\text{th}}} Mallows distance found in \insertCite{munk-czado-1998;textual}{JaB}
}
\details{
Add details here
}
