% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallows-distance.R
\name{mallows_distance_trimmed}
\alias{mallows_distance_trimmed}
\title{Trimmed Mallows Distance Between Two Vectors}
\usage{
mallows_distance_trimmed(vec1, vec2, p = 2, alpha = 0)
}
\arguments{
\item{vec1}{A numeric vector.}

\item{vec2}{A numeric vector of the same length as \code{vec1}.}

\item{p}{A number \eqn{\geq 1} for the \eqn{p^{\text{th}}} Mallows distance. Default is 2.}

\item{alpha}{a trimming parameter in \eqn{[0, 0.5)}. If \code{alpha}=0 (default), there is no trimming and result is just the Mallows distance.}
}
\value{
A numeric value representing the \eqn{p^{\text{th}}} Mallows distance between \code{vec1} and \code{vec2}.
}
\description{
Calculate the trimmed \eqn{p^{\text{th}}} Mallows distance of two vectors, also known as the trimmed Wasserstein-\eqn{p} distance.
}
\details{
Calculate the trimmed \eqn{p^{\text{th}}} Mallows distance of two vectors, also known as the trimmed Wasserstein-\eqn{p} distance between two samples of the same size.

Say \eqn{X_i \overset{\text{iid}}{\sim} F} and \eqn{Y_j \overset{\text{iid}}{\sim} G}.
Let \eqn{\hat{F}(x) = \sum_i\boldsymbol{1}(X_i \leq x)} and
\eqn{\hat{F}^{-1}(t) = \text{inf}\{x : \hat{F}(x) \leq t\}}.
Define \eqn{\hat{G}(y)} and \eqn{\hat{G}^{-1}(t)} similarly.

Let \eqn{\alpha\in[0,0/5)} be a trimming parameter.

Then the trimmed \eqn{p^{\text{th}}} Mallows distance between \eqn{\hat{F}(x)} and \eqn{\hat{G}(t)}  is

\eqn{\Psi_{\alpha, p}(\hat{F}, \hat{G}) = \left( \frac{1}{1-2\alpha}\int_{\alpha}^{1-\alpha} |\hat{F}^{-1}(t) - \hat{G}^{-1}(t) |^p dt \right)^{1/p}}
}
\examples{
data("karate")
set.seed(1)
boot.sample <- bootstrap_latent_space(karate, d = 2, B = 1)

orig.degree <- get_centrality(karate, "degree")
boot.degree <- get_centrality(boot.sample[[1]], "degree")

mallows_distance(orig.degree, boot.degree)
mallows_distance_trimmed(orig.degree, boot.degree, alpha = 1/34)
mallows_distance_trimmed(orig.degree, boot.degree, alpha = 2/34)

}
\references{
\insertRef{mallows-1972}{JaB}

\insertRef{munk-czado-1998}{JaB}
}
\seealso{
\code{\link[=mallows_distance]{mallows_distance()}}
}
