% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallows-distance-test.R
\name{mallows_equiv_test}
\alias{mallows_equiv_test}
\title{Equivalence Test of Two CDFs Using Mallows Distance}
\usage{
mallows_equiv_test(vec1, vec2, alpha, delta0, sig.level = 0.05)
}
\arguments{
\item{vec1}{A numeric vector.}

\item{vec2}{A numeric vector of the same length as \code{vec1}.}

\item{alpha}{Trimming parameter \eqn{\alpha \in(0, 0.5)} for trimmed Mallows' distance.}

\item{delta0}{Tolerance value for hypothesis test.}

\item{sig.level}{Significance level of the test. Default is 0.05.}
}
\value{
A list with the results of the test:
\itemize{
\item \code{dist.hat} is the estimated trimmed Mallows distance between the two distributions, \eqn{\Psi_{\alpha, 2}(\hat{F}, \hat{G})}.
\item \code{sd.hat} is the estimate of the standard deviation between the two distributions, \eqn{\hat{\sigma}_{\alpha}}.
\item \code{test.stat} is the test statistic.
\item \code{pval} is the p-value of the test.
\item \code{test.result} is  "REJECT NULL" if \code{pval}\eqn{\leq}\code{sig.level} and "FAIL TO REJECT NULL" otherwise.
\item \code{ci} is the upper (1-\code{sig.level}) confidence interval for the square of \code{dist.hat}
\item \code{alpha} is the trimming parameter used in the test. Sometimes this is slightly different than the input \code{alpha}. This test requires an integer \eqn{a} such that \eqn{\alpha = a/n}. This \code{alpha} is the smallest \eqn{\alpha} that makes this equation true. See \insertCite{munk-czado-1998;textual}{JaB} for details.
\item \code{delta0} is the tolerance parameter used for the test. This is the same as the input \code{delta0}.
}
}
\description{
Conduct the equivalence test of two CDFs using the \eqn{p^{\text{th}}} Mallows distance proposed by \insertCite{munk-czado-1998;textual}{JaB}
}
\details{
This is the equivalence test of two CDFs using the trimmed Mallows distance proposed by \insertCite{munk-czado-1998;textual}{JaB}.

Say \eqn{X_i \overset{iid}{\sim} F} and \eqn{Y_j \overset{iid}{\sim} G} for \eqn{i,j = 1, ..., n}
where \eqn{F} and \eqn{G} are continuous distribution functions.
Let \eqn{\hat{F}_n(x) = \frac{1}{n}\sum_{i=1}^n \boldsymbol{1}\{X_i \leq x\}}
denote the empirical cumulative distribution function (ECDF) of \eqn{X}, and
\eqn{\hat{F}^{-1}_n(t) = \inf\{x : F_n(x) \geq t\}} denote the quantile function.
Define \eqn{\hat{G}_m(y)} and \eqn{\hat{G}^{-1}_m(t)} similarly for \eqn{Y}.

The Trimmed \eqn{p^{th}} Mallows distance with trimming parameter \eqn{\alpha \in [0, 1/2)} is :
\eqn{\begin{equation}
\Psi_{\alpha, p}(F, G) =  \frac{1}{1-2\alpha} \left[ \int_{\alpha}^{1-\alpha} | F^{-1}(u) - G^{-1}(u) | ^p du \right]^{1/p}.
\end{equation} }

\insertCite{munk-czado-1998;textual}{JaB} conduct the equivalence test for some suitable \eqn{0<\Delta_0\in \mathbb{R}}:
\eqn{
\begin{equation}
H_0: \Psi_{\alpha, 2}(F, G) \geq \Delta_0 \quad \text{versus} \quad H_A:  \Psi_{\alpha, 2}(F, G) < \Delta_0
\end{equation} }

Then, a consistent level \eqn{\alpha^*} test for this hypothesis rejects \eqn{H_0} if,
\eqn{
\begin{equation}\label{eqn:MallowTest}
\left(\frac{nm}{n+m}\right)^{1/2} \frac{\Psi^2_{\alpha_{n \wedge m}, 2}(\hat{F}_n, \hat{G}_n) - \Delta_0^2}{\hat{\sigma}_{\alpha}(F, G)} \leq q_{\alpha^*}
\end{equation} }

where \eqn{q_{\alpha^*}} is the \eqn{\alpha^*} quantile of the standard normal distribution
and \eqn{\hat{\sigma}_{\alpha}} is a consistent estimator of the expected variance between \eqn{F} and \eqn{G}.
See \code{vignette("mallows-equiv-test")} or Appendix A of
\insertCite{munk-czado-1998;textual}{JaB} for the explicit expression of \eqn{\hat{\sigma}_{\alpha}}.

Currently, this function only runs when \eqn{n=m}. The theory still holds when \eqn{n\neq m}.

See \code{vignette("mallows-equiv-test")} for more details on the construction of this test.
}
\examples{
set.seed(2935)
X <- runif(500)
Y <- truncnorm::rtruncnorm(500, a = 0, b = 1, mean = 1/2, sd = 5)

# Plot the ECDFs
plot(ecdf(X), col = "red")
lines(ecdf(Y), col = "blue")

# Run the test for trimming parameter 0.05, tolerance value of 0.8 and
# significance level of 0.05.
test.result <- mallows_equiv_test(X, Y, alpha = 0.05, delta0 = 0.8)
test.result


}
