% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paul.revere.groups}
\alias{paul.revere.groups}
\title{Paul Revere Groups Network}
\format{
An undirected \code{igraph} graph object.

Vertex Attributes:
\itemize{
\item{"name": }{Name of the group}
\item{"members": }{list of people in \link[paul.revere]{paul.revere} who were known members of the group}
}

Edge Attributes:
\itemize{
\item{"weight": }{edge weight, people who were members of both groups}
\item{"shared.members": }{vector of length \code{weight} stating which people were members of both groups}
}
}
\source{
Data originally published by \insertCite{fischer-1994;textual}{JaB} and
electronically published on GitHub by \insertCite{healy-2017;textual}{JaB}.
Also found in \insertCite{han-2009;textual}{JaB}.
}
\usage{
paul.revere.groups
}
\description{
Association network of groups involved in the \link[paul.revere]{paul.revere}
network comprised of leaders from the American Revolution in the Boston area.
}
\details{
This network includes the seven groups that the 254 men of the \link[paul.revere]{paul.revere}
network were involved in:  the Masonic
lodge that met at the Green Dragon Tavern; the Loyal Nine, which was the
nucleus of the Sons of Liberty; the North Caucus that met at the Salutation Tavern; the
Long Room Club in Dassett Alley; the Boston Committee of Correspondence; the men who
are known to have participated in the Boston Tea Party; and Whig leaders on a Tory
Enemies List.

Nodes in this network are the groups. An edge exists between two groups if
at least one person was a member of both groups, and the edge weight is the number of people
who were members of both groups (e.g. Thomas Crafts and Henry Welles were both
members of St. Andrews Lodge and the Loyal Nine, so there is an edge of weight two
between St. Andrews Lodge and the Loyal Nine. There were no people in both the
Loyal Nine and the Long Room Club so there is not edge between them).
}
\examples{
data("paul.revere.groups")
set.seed(600)
plot(paul.revere.groups,
     edge.width = igraph::E(paul.revere.groups)$weight)

}
\references{
\insertAllCited{}
}
