% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap-latent-space.R
\name{ASE}
\alias{ASE}
\title{Calculate the Adjacency Spectral Embedding of a Matrix}
\usage{
ASE(graph, d)
}
\arguments{
\item{graph}{An igraph object with \eqn{n} nodes.}

\item{d}{number of latent dimensions to embed the node positions in.}
}
\value{
An \eqn{n}-by-\eqn{d} matrix where the \eqn{i^{th}} row of the matrix corresponds to node \eqn{i}'s position in the \eqn{d}-dimensional latent space.
Bootstrap samples will retain vertex names of the input `graph` but will loose all other vertex and edge attributes.
}
\description{
CITE Sussman et al. (2012) paper
}
\examples{
#ADD EXAMPLE ONCE DATA IS ADDED
}
