% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap-latent-space.R
\name{bootstrap_latent_space}
\alias{bootstrap_latent_space}
\title{Latent Space Bootstrap}
\usage{
bootstrap_latent_space(graph, d = 2, B = 1000, output.type = "igraph")
}
\arguments{
\item{graph}{An igraph object with $n$ nodes.}

\item{d}{Dimension of latent space. Recommended only 2 or 3 to maintain interpretability.}

\item{B}{number of bootstrap samples to calculate}

\item{output.type}{The class of object the resulting bootstrap networks should be.
The default is `igraph` which will make bootstrap samples of class "igraph". Note that for large \eqn{B}, this may not be an efficient use of storage space.
Other options include `matrix` which will return bootstrap samples as an \eqn{n}-by-\eqn{n} adjacency matrix,
`sparsematrix` which will return scarce matrices (package `Matrix` must be loaded).}
}
\value{
A list of length \eqn{B} where each element is an bootstrap sample.
Each element is of class `output.type`.
}
\description{
Add explanation of how latent space bootstrap works
}
\details{
Cite Sussman et al. (2012) paper for ASE
CITE Levin and Levina (2022) paper for bootstrap procedure.
Link to Matrix package
}
\examples{
# bootstrap_latent_space(paul.igraph, d=2, B=10, output.type = "matrix")

}
