% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-vertex.R
\name{bootstrap_vertex}
\alias{bootstrap_vertex}
\title{Vertex Bootstrap Sampling for networks}
\usage{
bootstrap_vertex(network, B = 1000, output.type = "igraph")
}
\arguments{
\item{network}{An igraph object representing the original network.}

\item{B}{An integer specifying the number of bootstrap samples to generate. Defaults to 1000.}

\item{output.type}{A character string specifying the format of the output networks. Options include \code{"igraph"}, \code{"matrix"}, \code{"dgCMatrix"}, or \code{"edgelist"}. Defaults to \code{"igraph"}. See \link[JaB:make_network_type]{make_network_type} for details.}
}
\value{
A list of bootstrap sample networks, each in the specified \code{output.type} format.
}
\description{
Generate bootstrap samples of a network by resampling vertices.
This function uses a modified version of the \link[snowboot:vertboot]{vertboot} from the \pkg{snowboot} package to allow flexible input and output network types.
}
\details{
Generates bootstrap samples of a network by resampling vertices.
This function uses a modified version of the \link[snowboot:vertboot]{vertboot} from the \pkg{snowboot} \insertCite{snowboot}{JaB} package to allow flexible input and output network types.

This procedure is first described in \insertCite{snijders-borgatti-1999;textual}{JaB}
and is outlined explicitly in \insertCite{chen-et-al-2018;textual}{JaB}.
}
\examples{
library(igraph)
data(karate)
set.seed(12)
boot.samp <- bootstrap_vertex(karate, B=1)

#plot comparison of original data and bootstrap sample
par(mfrow = c(1, 2))
#get the same positions in the original data and bootstrap samples
l <- igraph::layout_nicely(karate)
which.index <- NA
for(i in 1:gorder(boot.samp[[1]])){
  which.index[i] <- which(V(boot.samp[[1]])$name[i] == V(karate)$name)
}

plot(karate,
     layout = l,
     main = "Karate Data",
     vertex.label = NA)
plot(boot.samp[[1]],
     layout = l[which.index, ],
     main = "Karate Bootstrap Sample",
     vertex.label = NA,
     vertex.color = V(karate)$Faction[which.index ])


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{make_network_type}}, \code{\link[snowboot]{vertboot}}
}
