% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-vertex.R
\name{bootstrap_vertex}
\alias{bootstrap_vertex}
\title{Vertex Bootstrap Sampling for networks}
\usage{
bootstrap_vertex(network, B = 1000, output.type = "igraph")
}
\arguments{
\item{network}{An igraph object representing the original network.}

\item{B}{An integer specifying the number of bootstrap samples to generate. Defaults to 1000.}

\item{output.type}{A character string specifying the format of the output networks. Options include `"igraph"`, `"matrix"`, `"dgCMatrix"`, or `"edgelist"`. Defaults to `"igraph"`.}
}
\value{
A list of bootstrap sample networks, each in the specified `output.type` format.
}
\description{
Generates bootstrap samples of a network by resampling vertices. This function uses a modified version of the `vertboot` function from the `snowboot` package to allow flexible input and output network types.
}
\details{
The function performs vertex bootstrapping on the adjacency matrix of the input `network`, generating `B` bootstrap samples. Each sampled adjacency matrix is assigned row and column names based on the original vertex names. The function then converts the sampled matrices to the specified `output.type`.
CITE VERTEXT BOOTSTRAP PAPER, CITE THOMPSON ET AL, CITE SNOWBOOT
}
\examples{
# Create a sample igraph network
network <- igraph::make_ring(10)

# Generate bootstrap samples as igraph objects
boot_samples <- bootstrap_vertex(network, B = 100, output.type = "igraph")

# Generate bootstrap samples as adjacency matrices
boot_samples_matrices <- bootstrap_vertex(network, B = 100, output.type = "matrix")

}
\seealso{
\code{\link{make_network_type}}, \code{\link[snowboot]{vertboot}}
}
