% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-jackknife-after.R
\name{get_jackknife_after}
\alias{get_jackknife_after}
\title{Jackknife After-Bootstrap}
\usage{
get_jackknife_after(
  network,
  boot.result,
  quant,
  func.name,
  package.name = NULL,
  func.args = NULL
)
}
\arguments{
\item{network}{An igraph object representing the original network.}

\item{boot.result}{A list of bootstrap network samples, as produced by a \code{bootstrap_} function.}

\item{quant}{A numeric value of length 2 specifying the upper quantile used to flag influential nodes (e.g., \code{ 0.95}).}

\item{func.name}{A character string specifying the centrality function to be applied (e.g., \code{"degree"}, \code{"betweenness"}). See \code{\link{get_centrality}} for details}

\item{package.name}{(Optional) A character string specifying the package containing \code{func.name}. See \code{\link{get_centrality}} for details}

\item{func.args}{(Optional) A list of additional arguments for \code{func.name}.See \code{\link{get_centrality}} for details}
}
\value{
A data frame containing:
\itemize{
\item \code{Node_Number}: Numeric IDs of nodes.
\item \code{Node_Name}: Name of nodes.
\item \code{Orig_Stat}: The original centrality statistic of each node.
\item \code{Upper_Quantile}: Upper quantile of the jackknife-after-bootstrap distribution of centrality statistic for each node.
\item \code{Influential}: Logical indicating if each node is influential, i.e. is \code{Orig_Stat} greater than \code{Upper_Quantile}?
}
}
\description{
Detects influential nodes in a network with the jackknife-after-bootstrap algorithm.
}
\details{
This function performs a jackknife after-bootstrap procedure for each node in the network. It calculates the original centrality statistic for each node, then examines the bootstrap samples, excluding each node to generate a null distribution for that node’s centrality. Based on this distribution, it computes specified quantiles (default typically 0.025 and 0.975) to create confidence intervals.
}
\examples{
#add example later

}
