% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jab-lm.R
\name{jab_lm}
\alias{jab_lm}
\title{Jackknife-after-Bootstrap for Linear Regression}
\usage{
jab_lm(
  mod,
  stat = "rstudent",
  quant.lower = 0.05,
  quant.upper = 0.95,
  B = 3100,
  package.name = NULL,
  stat.args = NULL
)
}
\arguments{
\item{mod}{An \code{\link{lm}} object. Output from \code{\link[stats:lm]{stats::lm()}}.}

\item{stat}{A character string of the function name used to calculate the desired centrality statistic. The function must input an lm model object as is first argument and output a length \eqn{n} vector of the statistic of interest.}

\item{quant.lower}{A numeric between 0 and 1 used as the lower cutoff in the JaB algorithm. Default is 0.05. Must be smaller than \code{quant.upper}.}

\item{quant.upper}{A numeric between 0 and 1 used as the upper cutoff in the JaB algorithm. Default is 0.95. Must be larger than \code{quant.lower}.}

\item{B}{Number of bootstrap samples. Default is 3100}

\item{package.name}{A character string of the name of the package that \code{func.name} function is in. If left as \code{NULL} if the function will be called as loaded in the users environment.}

\item{stat.args}{A named list of additional arguments the \code{func.name} function may need beyond the \code{mod} object.}
}
\value{
A data frame with 5 columns.
\itemize{
\item{"row.ID": }{Row Number of the observation in the data set used for \code{mod}.}
\item{"lower": }{Lower quantile cutoff as determined by \code{quant.lower}. }
\item{"upper": }{Upper quantile cutoff as determined by \code{quant.upper}. }
\item{"orig": }{The original influence/outlier statistic calculated by \code{stat}.}
\item{"influential": }{Logical flagging if the observation is influential or not. TRUE if \code{orig} < \code{lower} or \code{orig} > \code{upper}. FALSE otherwise.}
}
}
\description{
Conduct Jackknife-after-Bootstrap (JaB) algorithm for linear regression \insertCite{martin-roberts-2013}{JaB}.
}
\details{
The JaB algorithm, proposed by \insertCite{martin-roberts-2013;textual}{JaB}
and further described by \insertCite{beyaztas-alin-2013;textual}{JaB}, detects
influential/outlier points in linear regression models. The algorithm is as follows:
\enumerate{
\item Let \eqn{\gamma} (\code{stat}) be the diagnostic statistic of interest. Fit the
model (\code{mod}) and calculate \eqn{\gamma_i} for \eqn{i=1,…,n}.

\item Construct \eqn{B} bootstrap samples, with replacement, from the original data set.
\item For \eqn{i = 1,…,n},
\itemize{
\item{3.1: }{Let \eqn{B_{(i)}} be the set of all bootstrap samples that did not contain data point \eqn{i}.}
\item{3.2: }{For each sample in \eqn{B_{(i)}}, fit the regression model then calculate the \eqn{n} values of \eqn{\gamma_{i, (b)}}. Aggregate them into one vector \eqn{\Gamma_i}.}
\item{3.1: }{Calculate suitable quantiles of \eqn{\Gamma_i} (\code{quant.lower} and \code{quant.upper}). If \eqn{\gamma_i} is outside of this range, flag point \eqn{i} as influential.}
}
}

The default for \code{quant.lower} and \code{quant.upper} is 0.05 and 0.95, respectively.
This means that if \eqn{\gamma_i} is in the center 90\% of \eqn{\Gamma_i}, it
will not be flagged as influential.

Some influence statistics, such as the likelihood distance from \insertCite{cook-1986;textual}{JaB},
are only positive values and large values imply influence. In these scenarios,
it is appropriate to set \code{quant.lower} to 0 and \code{quant.upper} to some suitable
quantile (say 0.90) so that point \eqn{i} is only flagged when \eqn{\gamma_i}
is in the upper quantiles of \eqn{\Gamma_i}.

As described in \insertCite{martin-roberts-2013;textual}{JaB}, to have approximately
1000 bootstrap samples in \eqn{B_{(i)}}, we need \eqn{Be^{1}\approx 3000} bootstrap
samples.

See \code{vignette("jab-regression")} for more details and examples.
}
\examples{
library(stats)
data("LifeCycleSavings")

mod <- lm(sr ~ ., data = LifeCycleSavings)

# JaB with DFFITS
result1 <- jab_lm(mod,
                  stat = "dffits",
                  quant.lower = 0.025,
                  quant.upper = 0.975,
                  B = 3100)
result1[result1$influential, ]


# define the likelihood distance as influence statistic (Cook, 1986)
infl_like <- function(mod){

  n <- length(mod$fitted.values)
  p <- length(mod$coefficients)

  ti <- rstudent(mod)
  h <- hatvalues(mod)

  p1 <- log( (n/(n-1)) * ((n-p-1) / (ti^2 +n-p-1)) )
  p2 <- ti^2 * (n-1) / (1-h) / (n-p-1)

  return(n*p1 + p2 - 1)
}

# JaB with Likelihood Distance
result2 <- jab_lm(mod,
                  stat = "infl_like",
                  quant.lower = 0.00,
                  quant.upper = 0.95,
                  B = 3100)
result2[result2$influential, ]
}
\references{
\insertAllCited{}
}
