% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{make_network_type}
\alias{make_network_type}
\title{Convert Network Objects to New Type}
\usage{
make_network_type(obj, input.type, output.type)
}
\arguments{
\item{obj}{A list of network objects, which must match the `input.type` specified.}

\item{input.type}{A character string specifying the input type of `obj`. Options include `"igraph"`, `"matrix"`, `"dgCMatrix"`, or `"edgelist"`.}

\item{output.type}{A character string specifying the desired output type. Options are `"igraph"`, `"matrix"`, `"dgCMatrix"`, or `"edgelist"`.}
}
\value{
A converted list of network objects in the specified `output.type` format.
}
\description{
Converts an input list of networks between various representations: `matrix`, `dgCMatrix`, `edgelist`, and `igraph`.
}
\details{
The function converts each network in `obj` to the specified `output.type`. Conversion logic is applied based on `input.type` and `output.type`.
- If `input.type` and `output.type` are the same, `obj` is returned as is.
- Conversions from adjacency matrices to igraph objects are handled via `igraph::graph_from_adjacency_matrix`.
- Conversions from igraph objects to edge lists are handled with `igraph::as_edgelist`.
- Conversions from igraph objects to adjacency matrices are handled with `igraph::as_adjacency_matrix`.

To use `"dgCMatrix"` for `input.type` or `output.type`, the package `Matrix` must be installed.
}
\examples{
# Convert list of adjacency matrices to igraph objects
adj_matrices <- replicate(3, matrix(c(0, 1, 1, 0), nrow = 2), simplify = FALSE)
make_network_type(adj_matrices, input.type = "matrix", output.type = "igraph")

# Convert list of igraph objects to sparse adjacency matrices
igraph_list <- lapply(adj_matrices, igraph::graph_from_adjacency_matrix)
make_network_type(igraph_list, input.type = "igraph", output.type = "dgCMatrix")

}
